+++ Source/Physics/ConstraintAtoms.cpp
@@ -12,12 +12,12 @@ AnchorAtom::AnchorAtom()
   mBodyR[0] = mBodyR[1] = Vec3::cZero;
 }
 
-Vec3Ref AnchorAtom::operator[](uint index)
+Vec3Ref AnchorAtom::operator[](size_t index)
 {
   return mBodyR[index];
 }
 
-Vec3Param AnchorAtom::operator[](uint index) const
+Vec3Param AnchorAtom::operator[](size_t index) const
 {
   return mBodyR[index];
 }
